% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-target.R
\name{ar_add_shots}
\alias{ar_add_shots}
\title{Add shots to the existing target plot.}
\usage{
ar_add_shots(data, colors = NULL)
}
\arguments{
\item{data}{archery data with shots to add.}

\item{colors}{vector of colors for new shots.}
}
\description{
Add shots to the existing target plot.
}
