% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped-mean.R
\name{ar_make_mean_grouped_data}
\alias{ar_make_mean_grouped_data}
\title{Create list with data grouped by N-shots for a given day}
\usage{
ar_make_mean_grouped_data(data, n = 6)
}
\arguments{
\item{data}{archery data for only one day}

\item{n}{size of the group}
}
\description{
Create list with data grouped by N-shots for a given day
}
