% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-target-compare.R
\name{archeryCompareTargetsModuleUI}
\alias{archeryCompareTargetsModuleUI}
\alias{archeryCompareTargetsModule}
\title{Module for comparing multiple targets}
\usage{
archeryCompareTargetsModuleUI(id)

archeryCompareTargetsModule(input, output, session, mainData, dates,
  MAX_PLOTS = 6)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{mainData}{}

\item{dates}{}

\item{MAX_PLOTS}{}
}
\description{
Module for comparing multiple targets
}
