% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-module.R
\name{archeryDataUI}
\alias{archeryDataUI}
\alias{archeryDataServer}
\title{Module for loading archery data}
\usage{
archeryDataUI(id)

archeryDataServer(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
Module for loading archery data
}
