% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-utils.R
\name{archeryDateModuleUI}
\alias{archeryDateModuleUI}
\alias{archeryDateModuleServer}
\title{Module for selecting dates}
\usage{
archeryDateModuleUI(id)

archeryDateModuleServer(input, output, session, mainData)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{mainData}{}
}
\description{
Module for selecting dates
}
