% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-utils.R
\name{archeryMiscellaneousPlotsUI}
\alias{archeryMiscellaneousPlotsUI}
\alias{archeryMiscellaneousPlotsServer}
\title{archeryMiscellaneousPlotsServer}
\usage{
archeryMiscellaneousPlotsUI(id)

archeryMiscellaneousPlotsServer(input, output, session, mainData,
  selectedDates)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{mainData}{}

\item{selectedDates}{}
}
\description{
archeryMiscellaneousPlotsServer
}
