% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex-hull.R
\name{plot_convex_hull_by}
\alias{plot_convex_hull_by}
\title{Plot target with convex hull.}
\usage{
plot_convex_hull_by(data, vars)
}
\arguments{
\item{data}{archery data.}

\item{vars}{vars to split by.}
}
\description{
Plot target with convex hull.
}
\examples{

data(archeryData)
plot_convex_hull_by(archeryData, "Date")

}
