% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plot_daily_total_points}
\alias{plot_daily_total_points}
\title{Plot total scores by day}
\usage{
plot_daily_total_points(data, addRollMean = TRUE, addRollMax = TRUE,
  n = 7, ...)
}
\arguments{
\item{data}{data.frame containing archery data.}

\item{...}{other parameters passed to \code{geom_line}.}
}
\description{
Plot total scores by day
}
\examples{

data(archeryData)
plot_daily_total_points(archeryData)

}
