% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot-groups-analysis.R
\name{plot_fnc_boxplot}
\alias{plot_fnc_boxplot}
\title{Create an functional boxplot for showing the distribution of results in single training session.}
\usage{
plot_fnc_boxplot(dtMatrix, bands = c(0, 0.5, 0.75, 1))
}
\arguments{
\item{dtMatrix}{data matrix. It should be a result of the \code{fnc_by_n_shots} or similar function.}

\item{bands}{a vector with bands.}
}
\description{
Create an functional boxplot for showing the distribution of results in single training session.
}
\examples{

data(archeryData)

dtMatrix12 <- fnc_by_n_shots(archeryData, 12)
plot_fnc_boxplot(dtMatrix12)

dtMatrix6 <- fnc_by_n_shots(archeryData, 6)
plot_fnc_boxplot(dtMatrix6)

dtMatrix3 <- fnc_by_n_shots(archeryData, 3)
plot_fnc_boxplot(dtMatrix3)

}
