% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plot_medians_polygon}
\alias{plot_medians_polygon}
\title{Plot daily medians.}
\usage{
plot_medians_polygon(data)
}
\arguments{
\item{data}{an archery data set}
}
\description{
Plot daily medians.
}
\examples{

data("archeryData")
plot_medians_polygon(archeryData)

}
