% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex-hull.R
\name{plot_scale_curves}
\alias{plot_scale_curves}
\title{Scale curves for archery data.}
\usage{
plot_scale_curves(data, vars)
}
\arguments{
\item{data}{archery data}

\item{vars}{variables to split by.}
}
\value{
ggplot2 object containing the scale curves.
}
\description{
Scale curves for archery data.
}
\examples{

data(archeryData)
plot_scale_curves(archeryData, "Date")

dates <- unique(archeryData$Date)
dates <- dates[c(1, length(dates))]
dt <- archeryData[archeryData[["Date"]] \%in\% dates,]
plot_scale_curves(dt, "Date") + 
  ggplot2::ggtitle("First day vs last")
  
}
