% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-target.R
\name{plot_target}
\alias{plot_target}
\title{Plot target}
\usage{
plot_target(target = ARCHERY_TARGETS$WA5_RING_40, setPars = FALSE)
}
\arguments{
\item{target}{a target object}
}
\description{
Plot target
}
\examples{

plot_target()

# set graphical parameters
plot_target(setPars = TRUE)

}
