% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-target.R
\name{plot_target_with_shots}
\alias{plot_target_with_shots}
\title{Plot target with shots}
\usage{
plot_target_with_shots(data)
}
\arguments{
\item{data}{archery data}
}
\description{
Plot target with shots
}
\examples{

data(archeryData)

dates <- unique(archeryData$Date)
dates <- dates[c(1, length(dates))]
firstDay <- archeryData[archeryData[["Date"]] == dates[1],]
lastDay <- archeryData[archeryData[["Date"]] == dates[2],]

oldPar <- par(
   no.readonly = TRUE, mfrow = c(1,2),
   mar = c(1,1,1,1), xpd = NA)
   
plot_target_with_shots(firstDay)
title(dates[1])
plot_target_with_shots(lastDay)
title(dates[2])
par(oldPar) 

}
