% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_archery_files}
\alias{read_archery_files}
\title{Read multiple files containing archery data.}
\usage{
read_archery_files(files)
}
\arguments{
\item{files}{a vector of csv to read.}
}
\value{
data.frame
}
\description{
Read multiple files containing archery data.
}
