% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchery.R
\name{run_archery_app}
\alias{run_archery_app}
\title{Start an aRchery}
\usage{
run_archery_app(port = 5005, host = "127.0.0.1",
  launch.browser = FALSE, ...)
}
\arguments{
\item{port}{The TCP port that the application should listen on.}

\item{host}{The IPv4 address that the application should listen on.}

\item{launch.browser}{If true, the system's default web browser will be launched automatically after the app is started.}

\item{...}{other parameters passed to \code{\link{runApp}}}
}
\description{
Start an aRchery
}
