% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot-groups-analysis.R
\name{fnc_by_n_shots}
\alias{fnc_by_n_shots}
\title{Summarise data for functional boxplot.}
\usage{
fnc_by_n_shots(data, n = 12)
}
\arguments{
\item{data}{}

\item{n}{}
}
\description{
Summarise data for functional boxplot.
}
\examples{

data(archeryData)
dtMatrix12 <- fnc_by_n_shots(archeryData, 12)

}
